+++ src/save_solution.h
@@ -3,7 +3,7 @@
 
 #include <string>
 
-void save_solution(std::string filename, int numPts, int numCells, double *q0, 
-                   double *q1, double *q2, double *q3, int *cellMap, int gam);
+void save_solution(std::string filename, int numPts, int numCells, double *q0,
+                   double *q1, double *q2, double *q3, int *cellMap, double gam);
 
 #endif
